# Conflicts:
#	utilities/car_system.py

+++ gym/f110_gym/envs/f110_env.py
@@ -105,7 +105,7 @@ class F110Env(gym.Env):
         
           
         # # Define action space
-        self.action_space = spaces.Box(low=np.array([-1.0, -1.0]), high=np.array([1.0, 1.0]), dtype=np.float32)
+        self.action_space = spaces.Box(low=np.array([-1.0, -1.0], dtype=np.float32), high=np.array([1.0, 1.0], dtype=np.float32), dtype=np.float32)
         # # Define observation space
         self.observation_space = spaces.Box(low=0, high=255, shape=(84, 84, 3), dtype=np.uint8)
